image-captioning-system/requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
clip-by-openai>=1.0
sentencepiece>=0.1.99

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0

# Web Framework and API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk[fastapi]>=1.30.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
tqdm>=4.65.0
wandb>=0.15.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Deployment
docker>=6.1.0
gunicorn>=21.0.0
